Move query dependency helpers into partial
