+++ src/CodeIndex/Cli/QueryCommandRunner.Dependencies.cs
@@ -1,6 +1,9 @@
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using CodeIndex.Database;
+using CodeIndex.Indexer;
+using CodeIndex.Models;
+using Microsoft.Data.Sqlite;
 
 namespace CodeIndex.Cli;
 
@@ -645,4 +648,950 @@ public static partial class QueryCommandRunner
             return CommandExitCodes.Success;
         }, cancellationToken: cancellationToken);
     }
+
+    private static bool TryExtractDepsFormat(string[] args, out string format, out string[] parseArgs, out string? error)
+    {
+        format = OutputFormatEdgeList;
+        error = null;
+        var rewritten = new List<string>(args.Length);
+        for (var i = 0; i < args.Length; i++)
+        {
+            var arg = args[i];
+            if (arg.StartsWith("--format=", StringComparison.Ordinal))
+            {
+                var rawFormat = arg["--format=".Length..];
+                if (!TryNormalizeDepsFormat(rawFormat, out format, out error))
+                {
+                    parseArgs = args;
+                    return false;
+                }
+                rewritten.Add(format == OutputFormatJsonGraph ? "--format=json" : "--format=text");
+                continue;
+            }
+
+            if (arg == "--format" && i + 1 < args.Length)
+            {
+                var rawFormat = args[++i];
+                if (!TryNormalizeDepsFormat(rawFormat, out format, out error))
+                {
+                    parseArgs = args;
+                    return false;
+                }
+                rewritten.Add("--format");
+                rewritten.Add(format == OutputFormatJsonGraph ? "json" : "text");
+                continue;
+            }
+
+            rewritten.Add(arg);
+        }
+
+        parseArgs = rewritten.ToArray();
+        return true;
+    }
+
+    private static bool TryNormalizeDepsFormat(string rawFormat, out string format, out string? error)
+    {
+        format = rawFormat.ToLowerInvariant();
+        error = null;
+        switch (format)
+        {
+            case OutputFormatText:
+            case OutputFormatJson:
+            case OutputFormatEdgeList:
+                format = OutputFormatEdgeList;
+                return true;
+            case OutputFormatDot:
+            case OutputFormatGraphMl:
+            case OutputFormatJsonGraph:
+                return true;
+            default:
+                error = $"Error: deps --format must be one of edgelist, dot, graphml, or json-graph; got '{ConsoleUi.FormatBoundedValue(rawFormat)}'.";
+                return false;
+        }
+    }
+
+    internal static List<FileDependencyResult> FilterCycleEdges(List<FileDependencyResult> results, out List<List<string>> cycles)
+        => FilterCycleEdges(results, out cycles, CancellationToken.None);
+
+    private static List<FileDependencyResult> FilterCycleEdges(IReadOnlyList<FileDependencyResult> results, out List<List<string>> cycles, CancellationToken cancellationToken)
+    {
+        cycles = FindDependencyCycles(results, cancellationToken);
+        if (cycles.Count == 0)
+            return [];
+        var cycleNodes = cycles.SelectMany(cycle => cycle).ToHashSet(StringComparer.Ordinal);
+        return results
+            .Where(edge => cycleNodes.Contains(edge.SourcePath) && cycleNodes.Contains(edge.TargetPath))
+            .ToList();
+    }
+
+    internal static List<List<string>> FindDependencyCycles(IReadOnlyList<FileDependencyResult> edges)
+        => FindDependencyCycles(edges, CancellationToken.None);
+
+    private static List<List<string>> FindDependencyCycles(IReadOnlyList<FileDependencyResult> edges, CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+        var adjacency = new Dictionary<string, List<string>>(StringComparer.Ordinal);
+        foreach (var edge in edges)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            if (!adjacency.TryGetValue(edge.SourcePath, out var targets))
+                adjacency[edge.SourcePath] = targets = [];
+            targets.Add(edge.TargetPath);
+            adjacency.TryAdd(edge.TargetPath, []);
+        }
+
+        var index = 0;
+        var stack = new Stack<string>();
+        var onStack = new HashSet<string>(StringComparer.Ordinal);
+        var indexes = new Dictionary<string, int>(StringComparer.Ordinal);
+        var lowLinks = new Dictionary<string, int>(StringComparer.Ordinal);
+        var cycles = new List<List<string>>();
+
+        void Visit(string node)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            indexes[node] = index;
+            lowLinks[node] = index;
+            index++;
+            stack.Push(node);
+            onStack.Add(node);
+
+            foreach (var target in adjacency[node])
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                if (!indexes.ContainsKey(target))
+                {
+                    Visit(target);
+                    lowLinks[node] = Math.Min(lowLinks[node], lowLinks[target]);
+                }
+                else if (onStack.Contains(target))
+                {
+                    lowLinks[node] = Math.Min(lowLinks[node], indexes[target]);
+                }
+            }
+
+            if (lowLinks[node] != indexes[node])
+                return;
+
+            var component = new List<string>();
+            string popped;
+            do
+            {
+                popped = stack.Pop();
+                onStack.Remove(popped);
+                component.Add(popped);
+            } while (!string.Equals(popped, node, StringComparison.Ordinal));
+
+            var selfCycle = component.Count == 1 && adjacency[component[0]].Contains(component[0], StringComparer.Ordinal);
+            if (component.Count > 1 || selfCycle)
+                cycles.Add(component.OrderBy(path => path, StringComparer.Ordinal).ToList());
+        }
+
+        foreach (var node in adjacency.Keys.OrderBy(path => path, StringComparer.Ordinal).ToList())
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            if (!indexes.ContainsKey(node))
+                Visit(node);
+        }
+
+        return cycles;
+    }
+
+    internal static JsonArray BuildDependencyCyclesJson(IReadOnlyList<List<string>> cycles)
+    {
+        var array = new JsonArray();
+        foreach (var cycle in cycles)
+        {
+            array.Add(new JsonObject
+            {
+                ["length"] = cycle.Count,
+                ["nodes"] = new JsonArray(cycle.Select(node => JsonValue.Create(node)).ToArray<JsonNode?>())
+            });
+        }
+        return array;
+    }
+
+    private sealed record DependencyCycleAnalysis(
+        List<FileDependencyResult> Edges,
+        List<List<string>> Cycles,
+        bool Truncated,
+        string TerminationReason,
+        string? TruncatedReason,
+        int CandidateEdgeCount,
+        int CandidateEdgeLimit,
+        int DisplayLimit,
+        string DetectionMode);
+
+    private static DependencyCycleAnalysis AnalyzeDependencyCycles(
+        IReadOnlyList<FileDependencyResult> candidateEdges,
+        int candidateEdgeLimit,
+        int candidateRowCount,
+        int displayLimit,
+        CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+        var edges = FilterCycleEdges(candidateEdges, out var allCycles, cancellationToken);
+        var cyclesTruncated = allCycles.Count > displayLimit;
+        var candidateTruncated = candidateRowCount > candidateEdgeLimit;
+        var cycles = allCycles.Take(displayLimit).ToList();
+        var cycleNodes = cycles.SelectMany(static cycle => cycle).ToHashSet(StringComparer.Ordinal);
+        var outputEdges = edges
+            .Where(edge => cycleNodes.Count == 0 || (cycleNodes.Contains(edge.SourcePath) && cycleNodes.Contains(edge.TargetPath)))
+            .Take(displayLimit)
+            .ToList();
+        var truncatedReason = candidateTruncated
+            ? "candidate_edge_limit"
+            : cyclesTruncated
+                ? "display_limit"
+                : null;
+        var terminationReason = truncatedReason switch
+        {
+            "candidate_edge_limit" => "candidate_limit_reached",
+            "display_limit" => "display_limit_reached",
+            _ => "completed",
+        };
+
+        return new DependencyCycleAnalysis(
+            outputEdges,
+            cycles,
+            candidateTruncated || cyclesTruncated,
+            terminationReason,
+            truncatedReason,
+            Math.Min(candidateRowCount, candidateEdgeLimit),
+            candidateEdgeLimit,
+            displayLimit,
+            DependencyCycleDetectionMode);
+    }
+
+    private static void AddDependencyCycleAnalysisJsonFields(JsonObject payload, DependencyCycleAnalysis analysis)
+        => AddDependencyCycleAnalysisJsonFields(
+            payload,
+            analysis.Truncated,
+            analysis.TerminationReason,
+            analysis.TruncatedReason,
+            analysis.CandidateEdgeCount,
+            analysis.CandidateEdgeLimit,
+            analysis.DisplayLimit,
+            analysis.DetectionMode);
+
+    internal static void AddDependencyCycleAnalysisJsonFields(
+        JsonObject payload,
+        bool truncated,
+        string terminationReason,
+        string? truncatedReason,
+        int candidateEdgeCount,
+        int candidateEdgeLimit,
+        int displayLimit,
+        string detectionMode,
+        JsonArray? nextStepFlags = null)
+    {
+        payload["truncated"] = truncated;
+        payload["termination_reason"] = terminationReason;
+        if (truncatedReason != null)
+            payload["truncated_reason"] = truncatedReason;
+        payload["candidate_edge_count"] = candidateEdgeCount;
+        payload["candidate_edge_limit"] = candidateEdgeLimit;
+        payload["cycle_detection_mode"] = detectionMode;
+        payload["cycle_result_scope"] = BuildDependencyCycleResultScope(truncatedReason);
+        payload["cycle_result_note"] = BuildDependencyCycleResultNote(truncatedReason);
+        payload["next_step_flags"] = nextStepFlags ?? BuildDependencyCycleNextStepFlagsJson(truncatedReason, candidateEdgeLimit, displayLimit);
+    }
+
+    private static string BuildDependencyCycleTruncationSummary(DependencyCycleAnalysis analysis)
+        => analysis.TruncatedReason == "display_limit"
+            ? $"partial: showing first {analysis.Cycles.Count} cycles"
+            : $"partial: candidate edge limit reached after {analysis.CandidateEdgeCount} candidate edges";
+
+    private static string BuildDependencyCycleTruncationWarning(DependencyCycleAnalysis analysis)
+    {
+        var nextSteps = BuildDependencyCycleNextStepFlags(
+            analysis.TruncatedReason,
+            analysis.CandidateEdgeLimit,
+            analysis.DisplayLimit);
+        var nextStepsText = nextSteps.Count == 0
+            ? string.Empty
+            : $" Next steps: {string.Join(", ", nextSteps)}.";
+        return "Warning: dependency cycle detection returned partial results "
+               + $"({BuildDependencyCycleTruncationSummary(analysis)}). "
+               + BuildDependencyCycleResultNote(analysis.TruncatedReason)
+               + nextStepsText;
+    }
+
+    private static string BuildDependencyCycleResultScope(string? truncatedReason)
+        => truncatedReason switch
+        {
+            "candidate_edge_limit" => "partial_candidate_edge_sample",
+            "display_limit" => "partial_display_limit",
+            _ => "complete_candidate_edges",
+        };
+
+    private static string BuildDependencyCycleResultNote(string? truncatedReason)
+        => truncatedReason switch
+        {
+            "candidate_edge_limit" => "Candidate edge limit reached before cdidx could prove cycle completeness; returned cycles are a bounded sample, not a complete or ranked cycle set.",
+            "display_limit" => "More cycles were detected than displayed; returned cycles are the first displayed cycles from the bounded candidate scan.",
+            _ => "Cycle detection completed for all candidate edges selected by the current filters.",
+        };
+
+    private static JsonArray BuildDependencyCycleNextStepFlagsJson(
+        string? truncatedReason,
+        int candidateEdgeLimit,
+        int displayLimit)
+        => new(BuildDependencyCycleNextStepFlags(truncatedReason, candidateEdgeLimit, displayLimit)
+            .Select(flag => JsonValue.Create(flag))
+            .ToArray<JsonNode?>());
+
+    internal static JsonArray BuildMcpDependencyCycleNextStepFlagsJson(
+        string? truncatedReason,
+        int candidateEdgeLimit,
+        int displayLimit)
+        => new(BuildMcpDependencyCycleNextStepFlags(truncatedReason, candidateEdgeLimit, displayLimit)
+            .Select(flag => JsonValue.Create(flag))
+            .ToArray<JsonNode?>());
+
+    private static List<string> BuildDependencyCycleNextStepFlags(
+        string? truncatedReason,
+        int candidateEdgeLimit,
+        int displayLimit)
+    {
+        var flags = new List<string>();
+        switch (truncatedReason)
+        {
+            case "candidate_edge_limit":
+                AddHigherLimitFlag(flags, candidateEdgeLimit);
+                flags.Add("--suppress-noise");
+                flags.Add("--symbol <name>");
+                flags.Add("--symbol-family <prefix>");
+                flags.Add("--path <narrower-glob>");
+                break;
+            case "display_limit":
+                AddHigherLimitFlag(flags, displayLimit);
+                flags.Add("--path <narrower-glob>");
+                break;
+        }
+
+        return flags;
+    }
+
+    private static List<string> BuildMcpDependencyCycleNextStepFlags(
+        string? truncatedReason,
+        int candidateEdgeLimit,
+        int displayLimit)
+    {
+        var flags = new List<string>();
+        switch (truncatedReason)
+        {
+            case "candidate_edge_limit":
+                AddHigherLimitArgument(flags, candidateEdgeLimit);
+                flags.Add("path=<narrower-glob>");
+                break;
+            case "display_limit":
+                AddHigherLimitArgument(flags, displayLimit);
+                flags.Add("path=<narrower-glob>");
+                break;
+        }
+
+        return flags;
+    }
+
+    private static void AddHigherLimitFlag(List<string> flags, int currentLimit)
+    {
+        var nextLimit = GetHigherDependencyCycleLimit(currentLimit);
+        if (nextLimit > currentLimit)
+            flags.Add($"--limit {nextLimit}");
+    }
+
+    private static void AddHigherLimitArgument(List<string> flags, int currentLimit)
+    {
+        var nextLimit = GetHigherDependencyCycleLimit(currentLimit);
+        if (nextLimit > currentLimit)
+            flags.Add($"limit={nextLimit}");
+    }
+
+    private static int GetHigherDependencyCycleLimit(int currentLimit)
+    {
+        var upperBound = NumericFlagUpperBounds.TryGetValue("--limit", out var maxLimit)
+            ? maxLimit
+            : int.MaxValue;
+        if (currentLimit >= upperBound)
+            return upperBound;
+
+        var doubled = currentLimit > upperBound / 2
+            ? upperBound
+            : currentLimit * 2;
+        return Math.Min(upperBound, Math.Max(currentLimit + 1, doubled));
+    }
+
+    private static DependencySymbolFilterResult ApplyDependencySymbolFilters(IReadOnlyList<FileDependencyResult> edges, QueryCommandOptions options)
+    {
+        var applied = options.DependencySuppressNoise || options.DependencySymbols.Count > 0 || options.DependencySymbolFamilies.Count > 0;
+        if (!applied)
+        {
+            var unchangedSymbolCount = edges.Sum(edge => SplitDependencySymbols(edge.Symbols).Count);
+            return new DependencySymbolFilterResult(
+                edges.ToList(),
+                new DependencySymbolFilterSummary(
+                    Applied: false,
+                    SuppressNoise: false,
+                    Symbols: [],
+                    SymbolFamilies: [],
+                    EdgesBefore: edges.Count,
+                    EdgesAfter: edges.Count,
+                    SymbolsBefore: unchangedSymbolCount,
+                    SymbolsAfter: unchangedSymbolCount));
+        }
+
+        var filteredEdges = new List<FileDependencyResult>(edges.Count);
+        var symbolsBefore = 0;
+        var symbolsAfter = 0;
+        foreach (var edge in edges)
+        {
+            var symbols = SplitDependencySymbols(edge.Symbols);
+            symbolsBefore += symbols.Count;
+            var keptSymbols = symbols
+                .Where(symbol => KeepDependencySymbol(symbol, options))
+                .ToList();
+            symbolsAfter += keptSymbols.Count;
+            if (keptSymbols.Count == 0)
+                continue;
+            filteredEdges.Add(CopyDependencyEdge(edge, string.Join(",", keptSymbols)));
+        }
+
+        return new DependencySymbolFilterResult(
+            filteredEdges,
+            new DependencySymbolFilterSummary(
+                Applied: true,
+                SuppressNoise: options.DependencySuppressNoise,
+                Symbols: options.DependencySymbols.ToArray(),
+                SymbolFamilies: options.DependencySymbolFamilies.ToArray(),
+                EdgesBefore: edges.Count,
+                EdgesAfter: filteredEdges.Count,
+                SymbolsBefore: symbolsBefore,
+                SymbolsAfter: symbolsAfter));
+    }
+
+    private static bool KeepDependencySymbol(string symbol, QueryCommandOptions options)
+    {
+        if (options.DependencySuppressNoise && DependencyNoiseProfile.IsNoiseSymbol(symbol))
+            return false;
+
+        var hasNameFilters = options.DependencySymbols.Count > 0 || options.DependencySymbolFamilies.Count > 0;
+        if (!hasNameFilters)
+            return true;
+
+        return options.DependencySymbols.Contains(symbol, StringComparer.Ordinal)
+               || options.DependencySymbolFamilies.Any(family => symbol.StartsWith(family, StringComparison.Ordinal));
+    }
+
+    private static List<string> SplitDependencySymbols(string symbols)
+        => symbols.Split(',', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries).ToList();
+
+    private static FileDependencyResult CopyDependencyEdge(FileDependencyResult edge, string symbols)
+        => new()
+        {
+            ResultKind = edge.ResultKind,
+            SourcePath = edge.SourcePath,
+            TargetPath = edge.TargetPath,
+            SourceDb = edge.SourceDb,
+            TargetDb = edge.TargetDb,
+            ReferenceCount = edge.ReferenceCount,
+            RankingScore = edge.RankingScore,
+            Symbols = symbols,
+        };
+
+    private static void AddDependencySchemaJsonFields(
+        JsonObject payload,
+        QueryCommandOptions options,
+        JsonSerializerOptions jsonOptions,
+        SqlGraphContractSignal sqlGraphSignal,
+        DependencySymbolFilterSummary symbolFilter)
+    {
+        payload["api_version"] = JsonOutputContract.ApiVersion;
+        payload["query_context"] = BuildQueryContextJson(options, jsonOptions);
+        AddSqlGraphContractJsonFields(payload, sqlGraphSignal);
+        AddDependencySymbolFilterJsonFields(payload, symbolFilter, jsonOptions);
+    }
+
+    private static void AddDependencySymbolFilterJsonFields(JsonObject payload, DependencySymbolFilterSummary symbolFilter, JsonSerializerOptions jsonOptions)
+    {
+        if (!symbolFilter.Applied)
+            return;
+
+        var filter = new JsonObject
+        {
+            ["suppress_noise"] = symbolFilter.SuppressNoise,
+            ["edges_before"] = symbolFilter.EdgesBefore,
+            ["edges_after"] = symbolFilter.EdgesAfter,
+            ["edges_removed"] = symbolFilter.EdgesBefore - symbolFilter.EdgesAfter,
+            ["symbols_before"] = symbolFilter.SymbolsBefore,
+            ["symbols_after"] = symbolFilter.SymbolsAfter,
+            ["symbols_removed"] = symbolFilter.SymbolsBefore - symbolFilter.SymbolsAfter,
+        };
+        if (symbolFilter.Symbols.Count > 0)
+            filter["symbol"] = JsonSerializer.SerializeToNode(symbolFilter.Symbols.ToList(), CliJsonSerializerContextFactory.Create(jsonOptions).ListString);
+        if (symbolFilter.SymbolFamilies.Count > 0)
+            filter["symbol_family"] = JsonSerializer.SerializeToNode(symbolFilter.SymbolFamilies.ToList(), CliJsonSerializerContextFactory.Create(jsonOptions).ListString);
+        payload["symbol_filter"] = filter;
+    }
+
+    private static JsonArray BuildDependencySymbolsJson(string symbols)
+        => new(SplitDependencySymbols(symbols).Select(symbol => JsonValue.Create(symbol)).ToArray<JsonNode?>());
+
+    private static bool DepsEmitsJson(QueryCommandOptions options, string depsFormat)
+        => depsFormat == OutputFormatJsonGraph || (options.Json && depsFormat == OutputFormatEdgeList);
+
+    private static int WriteDepsJsonPayload(JsonObject payload, QueryCommandOptions options, JsonSerializerOptions jsonOptions)
+        => WriteJsonPayloadWithOptionalByteLimit(
+            payload,
+            options,
+            jsonOptions,
+            "deps",
+            "deps",
+            "Use --summary-only, reduce --limit, or increase --max-json-bytes.");
+
+    private sealed record DependencySymbolFilterResult(List<FileDependencyResult> Edges, DependencySymbolFilterSummary Summary);
+
+    private sealed record DependencySymbolFilterSummary(
+        bool Applied,
+        bool SuppressNoise,
+        IReadOnlyList<string> Symbols,
+        IReadOnlyList<string> SymbolFamilies,
+        int EdgesBefore,
+        int EdgesAfter,
+        int SymbolsBefore,
+        int SymbolsAfter);
+
+    private static int WriteDependencyGraph(
+        IReadOnlyList<FileDependencyResult> edges,
+        string format,
+        JsonSerializerOptions jsonOptions,
+        DbReader reader,
+        QueryCommandOptions options,
+        SqlGraphContractSignal sqlGraphSignal,
+        DependencySymbolFilterSummary symbolFilter,
+        Action<JsonObject>? addExtraJsonFields = null)
+    {
+        switch (format)
+        {
+            case OutputFormatDot:
+                Console.WriteLine("digraph deps {");
+                foreach (var edge in edges)
+                    Console.WriteLine($"  \"{EscapeDot(edge.SourcePath)}\" -> \"{EscapeDot(edge.TargetPath)}\" [label=\"{edge.ReferenceCount}\"];");
+                Console.WriteLine("}");
+                return CommandExitCodes.Success;
+            case OutputFormatGraphMl:
+                Console.WriteLine("<?xml version=\"1.0\" encoding=\"UTF-8\"?>");
+                Console.WriteLine("<graphml xmlns=\"http://graphml.graphdrawing.org/xmlns\"><graph edgedefault=\"directed\">");
+                foreach (var node in edges.SelectMany(edge => new[] { edge.SourcePath, edge.TargetPath }).Distinct(StringComparer.Ordinal))
+                    Console.WriteLine($"<node id=\"{System.Security.SecurityElement.Escape(node)}\" />");
+                foreach (var edge in edges)
+                    Console.WriteLine($"<edge source=\"{System.Security.SecurityElement.Escape(edge.SourcePath)}\" target=\"{System.Security.SecurityElement.Escape(edge.TargetPath)}\"><data key=\"references\">{edge.ReferenceCount}</data></edge>");
+                Console.WriteLine("</graph></graphml>");
+                return CommandExitCodes.Success;
+            case OutputFormatJsonGraph:
+                return WriteDependencyJsonGraph(edges, jsonOptions, reader, options, sqlGraphSignal, symbolFilter, addExtraJsonFields);
+            default:
+                return CommandExitCodes.Success;
+        }
+    }
+
+    private static int WriteDependencyJsonGraph(
+        IReadOnlyList<FileDependencyResult> edges,
+        JsonSerializerOptions jsonOptions,
+        DbReader reader,
+        QueryCommandOptions options,
+        SqlGraphContractSignal sqlGraphSignal,
+        DependencySymbolFilterSummary symbolFilter,
+        Action<JsonObject>? addExtraJsonFields = null)
+    {
+        var seenNodes = new HashSet<string>(StringComparer.Ordinal);
+        var nodes = new List<string>();
+        foreach (var edge in edges)
+        {
+            if (seenNodes.Add(edge.SourcePath))
+                nodes.Add(edge.SourcePath);
+            if (seenNodes.Add(edge.TargetPath))
+                nodes.Add(edge.TargetPath);
+        }
+
+        var payload = new JsonObject { ["count"] = edges.Count };
+        if (options.SummaryOnly)
+            payload["summary_only"] = true;
+        else
+        {
+            payload["nodes"] = new JsonArray(nodes.Select(node => (JsonNode?)new JsonObject { ["id"] = node }).ToArray());
+            payload["edges"] = new JsonArray(edges.Select(edge => (JsonNode?)new JsonObject
+            {
+                ["source"] = edge.SourcePath,
+                ["target"] = edge.TargetPath,
+                ["reference_count"] = edge.ReferenceCount,
+                ["ranking_score"] = edge.RankingScore,
+                ["symbols"] = BuildDependencySymbolsJson(edge.Symbols),
+            }).ToArray());
+        }
+        addExtraJsonFields?.Invoke(payload);
+        AddDependencySchemaJsonFields(payload, options, jsonOptions, sqlGraphSignal, symbolFilter);
+        AddFreshnessHint(payload, reader);
+        return WriteDepsJsonPayload(payload, options, jsonOptions);
+    }
+
+    private static string EscapeDot(string value) => value.Replace("\\", "\\\\", StringComparison.Ordinal).Replace("\"", "\\\"", StringComparison.Ordinal);
+
+    internal static int GetDependencyCycleGraphLimit(int displayLimit)
+    {
+        var requestedLimit = Math.Max(displayLimit, DefaultDependencyCycleGraphLimit);
+        return NumericFlagUpperBounds.TryGetValue("--limit", out var maxLimit)
+            ? Math.Min(requestedLimit, maxLimit)
+            : requestedLimit;
+    }
+
+    private static List<FileDependencyResult> GetWorkspaceFileDependencies(DbReader primaryReader, QueryCommandOptions options, bool reverse, int limit, CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+        var results = primaryReader.GetFileDependencies(limit, options.Lang, options.PathPatterns, options.ExcludePaths, options.ExcludeTests, reverse, cancellationToken);
+        cancellationToken.ThrowIfCancellationRequested();
+        if (options.WorkspaceDbPaths.Count == 0)
+            return results;
+
+        var memberDbs = BuildWorkspaceDependencyDatabaseList(options);
+        var primaryDb = memberDbs[0];
+        TagFileDependencyResults(results, primaryDb);
+        foreach (var normalizedDbPath in memberDbs.Skip(1))
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            using var db = new DbContext(normalizedDbPath, cancellationToken);
+            db.TryMigrateForRead();
+            var reader = new DbReader(db) { IncludeGenerated = primaryReader.IncludeGenerated };
+            var memberResults = reader.GetFileDependencies(limit, options.Lang, options.PathPatterns, options.ExcludePaths, options.ExcludeTests, reverse, cancellationToken);
+            TagFileDependencyResults(memberResults, normalizedDbPath);
+            results.AddRange(memberResults);
+        }
+
+        foreach (var sourceDb in memberDbs)
+            foreach (var targetDb in memberDbs)
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                if (string.Equals(sourceDb, targetDb, StringComparison.Ordinal))
+                    continue;
+                results.AddRange(GetCrossDatabaseFileDependencies(sourceDb, targetDb, options, reverse, limit));
+            }
+
+        return results
+            .OrderByDescending(result => result.RankingScore)
+            .ThenByDescending(result => result.ReferenceCount)
+            .ThenBy(result => result.SourceDb, StringComparer.Ordinal)
+            .ThenBy(result => result.SourcePath, StringComparer.Ordinal)
+            .ThenBy(result => result.TargetDb, StringComparer.Ordinal)
+            .ThenBy(result => result.TargetPath, StringComparer.Ordinal)
+            .Take(limit)
+            .ToList();
+    }
+
+    private static List<FileDependencyResult> GetWorkspaceFileDependencyCycleCandidates(
+        DbReader primaryReader,
+        QueryCommandOptions options,
+        bool reverse,
+        int limit,
+        out int candidateRowCount,
+        CancellationToken cancellationToken)
+    {
+        candidateRowCount = 0;
+        var results = primaryReader.GetFileDependencyCycleCandidates(
+            limit,
+            out var primaryCandidateRows,
+            options.Lang,
+            options.PathPatterns,
+            options.ExcludePaths,
+            options.ExcludeTests,
+            reverse,
+            cancellationToken);
+        candidateRowCount += primaryCandidateRows;
+        if (options.WorkspaceDbPaths.Count == 0)
+            return results.Take(limit).ToList();
+
+        var memberDbs = BuildWorkspaceDependencyDatabaseList(options);
+        var primaryDb = memberDbs[0];
+        TagFileDependencyResults(results, primaryDb);
+        if (results.Count >= limit)
+            return results.Take(limit).ToList();
+        foreach (var normalizedDbPath in memberDbs.Skip(1))
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            using var db = new DbContext(normalizedDbPath, cancellationToken);
+            db.TryMigrateForRead();
+            var reader = new DbReader(db) { IncludeGenerated = primaryReader.IncludeGenerated };
+            var memberResults = reader.GetFileDependencyCycleCandidates(
+                limit,
+                out var memberCandidateRows,
+                options.Lang,
+                options.PathPatterns,
+                options.ExcludePaths,
+                options.ExcludeTests,
+                reverse,
+                cancellationToken);
+            candidateRowCount += memberCandidateRows;
+            TagFileDependencyResults(memberResults, normalizedDbPath);
+            results.AddRange(memberResults);
+            if (results.Count >= limit)
+                return results.Take(limit).ToList();
+        }
+
+        foreach (var sourceDb in memberDbs)
+            foreach (var targetDb in memberDbs)
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                if (string.Equals(sourceDb, targetDb, StringComparison.Ordinal))
+                    continue;
+                var crossDbResults = GetCrossDatabaseFileDependencies(sourceDb, targetDb, options, reverse, limit);
+                candidateRowCount += crossDbResults.Count;
+                results.AddRange(crossDbResults);
+                if (results.Count >= limit)
+                    return results.Take(limit).ToList();
+            }
+
+        return results.Take(limit).ToList();
+    }
+
+    internal static List<string> BuildWorkspaceDependencyDatabaseList(QueryCommandOptions options)
+    {
+        var primaryDb = Path.GetFullPath(DbPathResolver.NormalizeDbPath(options.DbPath));
+        var comparer = PathCasing.IsIgnoreCase(primaryDb)
+            ? StringComparer.OrdinalIgnoreCase
+            : StringComparer.Ordinal;
+        return options.WorkspaceDbPaths
+            .Select(path => Path.GetFullPath(DbPathResolver.NormalizeDbPath(path)))
+            .Prepend(primaryDb)
+            .Distinct(comparer)
+            .ToList();
+    }
+
+    private static bool TryWriteWorkspaceDependencyFanOutError(QueryCommandOptions options)
+    {
+        if (options.WorkspaceDbPaths.Count == 0)
+            return false;
+
+        var memberDbs = BuildWorkspaceDependencyDatabaseList(options);
+        var pairCount = memberDbs.Count * (memberDbs.Count - 1);
+        if (memberDbs.Count <= MaxWorkspaceDependencyDatabaseCount &&
+            pairCount <= MaxWorkspaceDependencyDatabasePairCount)
+            return false;
+
+        var maxAdditional = MaxWorkspaceDependencyDatabaseCount - 1;
+        var additionalCount = Math.Max(0, memberDbs.Count - 1);
+        CommandErrorWriter.WriteStderr($"Error: deps --workspace-db accepts at most {maxAdditional} distinct additional databases ({MaxWorkspaceDependencyDatabaseCount} total including --db), which is {MaxWorkspaceDependencyDatabasePairCount} ordered cross-database pairs; got {additionalCount} additional ({memberDbs.Count} total, {pairCount} pairs).");
+        CommandErrorWriter.WriteStderr("Hint: pass fewer --workspace-db values or run deps separately for smaller workspace member groups.");
+        return true;
+    }
+
+    private static bool TryWriteInvalidWorkspaceDependencyDatabaseError(QueryCommandOptions options, out int exitCode)
+    {
+        exitCode = CommandExitCodes.Success;
+        if (options.WorkspaceDbPaths.Count == 0)
+            return false;
+
+        foreach (var dbPath in BuildWorkspaceDependencyDatabaseList(options).Skip(1))
+        {
+            if (DbContext.TryValidateExistingCodeIndexDb(
+                    dbPath,
+                    requireWritable: false,
+                    requireSupportedUserVersion: true,
+                    out var validationMessage,
+                    out var isNotFound,
+                    out var isSchemaTooNew))
+                continue;
+
+            var errorCode = isNotFound
+                ? CommandErrorCodes.DbNotFound
+                : isSchemaTooNew
+                    ? CommandErrorCodes.SchemaTooNew
+                    : CommandErrorCodes.DbError;
+            CommandErrorWriter.WriteStderr($"Error [{errorCode}]: attached workspace database cannot be used for cross-database dependency query: {validationMessage}");
+            CommandErrorWriter.WriteStderr(isNotFound
+                ? "Hint: pass an existing CodeIndex database to `--workspace-db`, or run `cdidx index <workspacePath>` for that workspace member first."
+                : isSchemaTooNew
+                    ? "Hint: run the query with a current cdidx binary, or rebuild that workspace member database with this cdidx version before using `--workspace-db`."
+                    : "Hint: pass only CodeIndex databases created by `cdidx index` to `--workspace-db`; remove stale, empty, or unrelated SQLite files from the workspace database list.");
+            exitCode = CommandExitCodes.DatabaseError;
+            return true;
+        }
+
+        return false;
+    }
+
+    private static List<FileDependencyResult> GetCrossDatabaseFileDependencies(string sourceDbPath, string targetDbPath, QueryCommandOptions options, bool reverse, int limit)
+    {
+        using var sourceDb = new DbContext(sourceDbPath);
+        sourceDb.TryMigrateForRead();
+        var connection = sourceDb.Connection;
+        AttachCrossDatabaseTarget(connection, targetDbPath);
+
+        using var cmd = connection.CreateCommand();
+        var sourcePathExpr = reverse ? "dst.path" : "src.path";
+        var targetPathExpr = reverse ? "src.path" : "dst.path";
+        cmd.CommandText = $@"
+            WITH edges AS (
+            SELECT {sourcePathExpr} AS source_path,
+                   {targetPathExpr} AS target_path,
+                   r.symbol_name
+            FROM symbol_references r
+            JOIN files src ON src.id = r.file_id
+            JOIN targetdb.symbols s ON s.name = r.symbol_name
+            JOIN targetdb.files dst ON dst.id = s.file_id
+            WHERE 1 = 1";
+        if (options.Lang != null)
+        {
+            cmd.CommandText += " AND src.lang = @lang AND dst.lang = @lang";
+            SqliteCommandPolicy.Add(cmd, "@lang", options.Lang);
+        }
+        AddCrossDatabasePathFilters(cmd, "src", options.PathPatterns, include: !reverse);
+        AddCrossDatabasePathFilters(cmd, "dst", options.PathPatterns, include: reverse);
+        AddCrossDatabaseExcludeFilters(cmd, "src", options.ExcludePaths, include: !reverse);
+        AddCrossDatabaseExcludeFilters(cmd, "dst", options.ExcludePaths, include: reverse);
+        if (options.ExcludeTests)
+            cmd.CommandText += reverse
+                ? $" AND NOT {BuildCrossDatabaseTestPathCondition("dst")}"
+                : $" AND NOT {BuildCrossDatabaseTestPathCondition("src")}";
+        cmd.CommandText += @"
+            ),
+            edge_totals AS (
+                SELECT source_path,
+                       target_path,
+                       COUNT(*) AS reference_count
+                FROM edges
+                GROUP BY source_path, target_path
+            ),
+            distinct_edge_symbols AS (
+                SELECT DISTINCT source_path, target_path, symbol_name
+                FROM edges
+            ),
+            ranked_edge_symbols AS (
+                SELECT source_path,
+                       target_path,
+                       symbol_name,
+                       ROW_NUMBER() OVER (PARTITION BY source_path, target_path ORDER BY symbol_name) AS symbol_rank
+                FROM distinct_edge_symbols
+            )
+            SELECT edge_totals.source_path,
+                   edge_totals.target_path,
+                   edge_totals.reference_count,
+                   COALESCE(GROUP_CONCAT(CASE WHEN ranked_edge_symbols.symbol_rank <= @symbolSampleLimit THEN ranked_edge_symbols.symbol_name END), '') AS symbols
+            FROM edge_totals
+            LEFT JOIN ranked_edge_symbols
+              ON ranked_edge_symbols.source_path = edge_totals.source_path
+             AND ranked_edge_symbols.target_path = edge_totals.target_path
+            GROUP BY edge_totals.source_path, edge_totals.target_path, edge_totals.reference_count
+            ORDER BY edge_totals.reference_count DESC, edge_totals.source_path, edge_totals.target_path
+            LIMIT @limit";
+        SqliteCommandPolicy.Add(cmd, "@limit", DependencyNoiseProfile.GetRankingCandidateLimit(limit));
+        SqliteCommandPolicy.Add(cmd, "@symbolSampleLimit", DbReader.DependencySymbolSampleLimit);
+
+        var results = new List<FileDependencyResult>();
+        using var reader = cmd.ExecuteReader();
+        while (reader.Read())
+        {
+            results.Add(new FileDependencyResult
+            {
+                SourcePath = reader.GetString(0),
+                TargetPath = reader.GetString(1),
+                SourceDb = reverse ? targetDbPath : sourceDbPath,
+                TargetDb = reverse ? sourceDbPath : targetDbPath,
+                ReferenceCount = reader.GetInt32(2),
+                Symbols = reader.IsDBNull(3) ? string.Empty : reader.GetString(3),
+            });
+        }
+        foreach (var result in results)
+            result.RankingScore = DependencyNoiseProfile.ComputeRankingScore(result.ReferenceCount, result.Symbols);
+
+        return results
+            .OrderByDescending(result => result.RankingScore)
+            .ThenByDescending(result => result.ReferenceCount)
+            .ThenBy(result => result.SourceDb, StringComparer.Ordinal)
+            .ThenBy(result => result.SourcePath, StringComparer.Ordinal)
+            .ThenBy(result => result.TargetDb, StringComparer.Ordinal)
+            .ThenBy(result => result.TargetPath, StringComparer.Ordinal)
+            .Take(limit)
+            .ToList();
+    }
+
+    private static void AttachCrossDatabaseTarget(SqliteConnection connection, string targetDbPath)
+    {
+        try
+        {
+            AttachCrossDatabaseTargetCore(connection, targetDbPath);
+        }
+        catch (SqliteException) when (!SqliteFileUri.StartsWithFileScheme(targetDbPath) && File.Exists(LongPath.EnsureWindowsPrefix(targetDbPath)))
+        {
+            AttachCrossDatabaseTargetCore(connection, DbContext.ToReadOnlyUri(targetDbPath));
+        }
+    }
+
+    private static void AttachCrossDatabaseTargetCore(SqliteConnection connection, string targetDbPath)
+    {
+        using var attach = connection.CreateCommand();
+        attach.CommandText = "ATTACH DATABASE @targetDb AS targetdb";
+        attach.Parameters.Add("@targetDb", SqliteType.Text).Value = targetDbPath;
+        attach.ExecuteNonQuery();
+    }
+
+    private static void AddCrossDatabasePathFilters(SqliteCommand cmd, string alias, IReadOnlyList<string> patterns, bool include)
+    {
+        if (!include || patterns.Count == 0)
+            return;
+        SqliteDynamicSql.EnsureParameterBudget(patterns.Count, "cross-database path filters");
+        var parts = new List<string>(patterns.Count);
+        for (var i = 0; i < patterns.Count; i++)
+        {
+            var name = SqliteDynamicSql.BuildParameterName($"crossPath{alias}", i);
+            parts.Add($"{alias}.path LIKE {name} ESCAPE '\\'");
+            cmd.Parameters.Add(name, SqliteType.Text).Value = CrossDatabaseGlobToLikePattern(patterns[i]);
+        }
+        cmd.CommandText += " AND (" + string.Join(" OR ", parts) + ")";
+    }
+
+    private static void AddCrossDatabaseExcludeFilters(SqliteCommand cmd, string alias, IReadOnlyList<string> patterns, bool include)
+    {
+        if (!include || patterns.Count == 0)
+            return;
+        SqliteDynamicSql.EnsureParameterBudget(patterns.Count, "cross-database exclude path filters");
+        for (var i = 0; i < patterns.Count; i++)
+        {
+            var name = SqliteDynamicSql.BuildParameterName($"crossExclude{alias}", i);
+            cmd.CommandText += $" AND {alias}.path NOT LIKE {name} ESCAPE '\\'";
+            cmd.Parameters.Add(name, SqliteType.Text).Value = CrossDatabaseGlobToLikePattern(patterns[i]);
+        }
+    }
+
+    internal static string BuildCrossDatabaseTestPathConditionForTesting(string alias)
+        => BuildCrossDatabaseTestPathCondition(alias);
+
+    private static string BuildCrossDatabaseTestPathCondition(string alias)
+        => DbReader.TestPathCondition.Replace("f.path", $"{alias}.path", StringComparison.Ordinal);
+
+    private static string CrossDatabaseGlobToLikePattern(string pattern)
+    {
+        var builder = new System.Text.StringBuilder(pattern.Length);
+        foreach (var ch in pattern)
+        {
+            builder.Append(ch switch
+            {
+                '*' => '%',
+                '?' => '_',
+                '%' => "\\%",
+                '_' => "\\_",
+                '\\' => "\\\\",
+                _ => ch,
+            });
+        }
+        return builder.ToString();
+    }
+
+    private static void TagFileDependencyResults(IEnumerable<FileDependencyResult> results, string dbPath)
+    {
+        foreach (var result in results)
+        {
+            result.SourceDb = dbPath;
+            result.TargetDb = dbPath;
+        }
+    }
+
 }
